src/requirements.txt
pigpio
python-dotenv
requests
yaml-config-parser